Added A Wave Spawning System
So I added this wave spawner system following a youtube tutorial and plugging in my enemy classes. It works ok, for some reason I have to have at least 10 enemies in each wave which kind of sucks. But this is what I got so far. I think later I'll try and fix the minimum 10 enemies problem